cross_leverage_limit may decimals (#20)
* cross_leverage_limit may decimals

+++ GateIo.Net/Objects/Models/GateIoPositionUpdate.cs
@@ -97,6 +97,6 @@ namespace GateIo.Net.Objects.Models
         /// Cross margin leverage
         /// </summary>
         [JsonPropertyName("cross_leverage_limit")]
-        public int? CrossLeverageLimit { get; set; }
+        public decimal? CrossLeverageLimit { get; set; }
     }
 }
